requirements.txt
asgiref==3.8.1
cloudinary==1.40.0
dj-rest-auth==6.0.0
Django==5.0.6
django-cloudinary-storage==0.3.0
django-filter==24.2
djangorestframework==3.15.2
pillow==10.4.0
sqlparse==0.5.0
